feat(readOnly): support read only datahub deployments

+++ metadata-io/src/main/java/com/linkedin/metadata/graph/elastic/ESGraphWriteDAO.java
@@ -27,6 +27,11 @@ public class ESGraphWriteDAO {
   private final ESBulkProcessor bulkProcessor;
   private final int numRetries;
   private final GraphQueryConfiguration graphQueryConfiguration;
+  private boolean canWrite = true;
+
+  public void setWritable(boolean writable) {
+    canWrite = writable;
+  }
 
   /**
    * Updates or inserts the given search document.
@@ -35,6 +40,9 @@ public class ESGraphWriteDAO {
    * @param docId the ID of the document
    */
   public void upsertDocument(@Nonnull String docId, @Nonnull String document) {
+    if (!canWrite) {
+      return;
+    }
     final UpdateRequest updateRequest =
         new UpdateRequest(indexConvention.getIndexName(INDEX_NAME), docId)
             .detectNoop(false)
@@ -50,20 +58,31 @@ public class ESGraphWriteDAO {
    * @param docId the ID of the document
    */
   public void deleteDocument(@Nonnull String docId) {
+    if (!canWrite) {
+      return;
+    }
     final DeleteRequest deleteRequest =
         new DeleteRequest(indexConvention.getIndexName(INDEX_NAME)).id(docId);
     bulkProcessor.add(deleteRequest);
   }
 
+  @Nullable
   public BulkByScrollResponse deleteByQuery(
       @Nonnull final OperationContext opContext, @Nonnull final GraphFilters graphFilters) {
+    if (!canWrite) {
+      return null;
+    }
     return deleteByQuery(opContext, graphFilters, null);
   }
 
+  @Nullable
   public BulkByScrollResponse deleteByQuery(
       @Nonnull final OperationContext opContext,
       @Nonnull final GraphFilters graphFilters,
       String lifecycleOwner) {
+    if (!canWrite) {
+      return null;
+    }
     BoolQueryBuilder finalQuery =
         buildQuery(opContext, graphQueryConfiguration, graphFilters, lifecycleOwner);
 
@@ -75,6 +94,9 @@ public class ESGraphWriteDAO {
   @Nullable
   public BulkByScrollResponse updateByQuery(
       @Nonnull Script script, @Nonnull final QueryBuilder query) {
+    if (!canWrite) {
+      return null;
+    }
     return bulkProcessor
         .updateByQuery(script, query, indexConvention.getIndexName(INDEX_NAME))
         .orElse(null);
